pyproject.toml
[tool.poetry]
name = "refactortest"
version = "0.1.0"
description = "hi nate"
authors = ["Evan Green <emg5555@psu.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
